+++ webapp/php/app/routes.php
@@ -7,267 +7,15 @@ use Psr\Log\LoggerInterface;
 use Fig\Http\Message\StatusCodeInterface;
 use League\Csv\Reader;
 use Slim\App;
+use App\Domain\Chair;
 
 const EXEC_SUCCESS = 127;
 const DIR_ROOT = __DIR__ . '/../..';
-
-class Chair
-{
-    public function getId(): ?int
-    {
-        return (int)$this->id;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function getDescription(): ?string
-    {
-        return $this->description;
-    }
-
-    public function getThumbnail(): ?string
-    {
-        return $this->thumbnail;
-    }
-
-    public function getPrice(): ?int
-    {
-        return (int)$this->price;
-    }
-
-    public function getHeight(): ?int
-    {
-        return (int)$this->height;
-    }
-
-    public function getWidth(): ?int
-    {
-        return (int)$this->width;
-    }
-
-    public function getDepth(): ?int
-    {
-        return (int)$this->depth;
-    }
-
-    public function getColor(): ?string
-    {
-        return $this->color;
-    }
-
-    public function getFeatures(): ?string
-    {
-        return $this->features;
-    }
-
-    public function getKind(): ?string
-    {
-        return $this->kind;
-    }
-
-    public function getPopularity(): ?int
-    {
-        return (int)$this->popularity;
-    }
-
-    public function getStock(): ?int
-    {
-        return (int)$this->stock;
-    }
-
-    public function toArray()
-    {
-        return [
-            'id' => $this->getId(),
-            'name' => $this->getName(),
-            'description' => $this->getDescription(),
-            'thumbnail' => $this->getThumbnail(),
-            'price' => $this->getPrice(),
-            'height' => $this->getHeight(),
-            'width' => $this->getWidth(),
-            'depth' => $this->getDepth(),
-            'color' => $this->getColor(),
-            'features' => $this->getFeatures(),
-            'kind' => $this->getKind(),
-            'popularity' => $this->getPopularity(),
-            'stock' => $this->getStock(),
-        ];
-    }
-}
-
-class Range
-{
-    public ?int $id;
-    public ?int $min;
-    public ?int $max;
-
-    public function __construct(
-        int $id = null,
-        int $min = null,
-        int $max = null
-    ) {
-        $this->id = $id;
-        $this->min = $min;
-        $this->max = $max;
-    }
-
-    public static function unmarshal(array $json) {
-        return new Range(
-            $json['id'] ?? null,
-            $json['min'] ?? null,
-            $json['max'] ?? null,
-        );
-    }
-}
-
-class RangeCondition
-{
-    public ?string $prefix;
-    public ?string $suffix;
-    /** @var Range[] */
-    public array $ranges;
-
-    public function __construct(
-        string $prefix = null,
-        string $suffix = null,
-        array $ranges = []
-    ) {
-        $this->prefix = $prefix;
-        $this->suffix = $suffix;
-        $this->ranges = $ranges;
-    }
-
-    public static function unmarshal(array $json): RangeCondition
-    {
-        return new RangeCondition(
-            $json['prefix'] ?? null,
-            $json['suffix'] ?? null,
-            array_map(Range::class . '::unmarshal', $json['ranges'] ?? [])
-        );
-    }
-}
-
-class ChairSearchCondition
-{
-    public ?RangeCondition $width;
-    public ?RangeCondition $height;
-    public ?RangeCondition $depth;
-    public ?RangeCondition $price;
-    public ?RangeCondition $color;
-    public ?RangeCondition $feature;
-    public ?RangeCondition $kind;
-
-    public function __construct(
-        RangeCondition $width = null,
-        RangeCondition $height = null,
-        RangeCondition $depth = null,
-        RangeCondition $price = null,
-        RangeCondition $color = null,
-        RangeCondition $feature = null,
-        RangeCondition $kind = null
-    ) {
-        $this->width = $width;
-        $this->height = $height;
-        $this->depth = $depth;
-        $this->price = $price;
-        $this->color = $color;
-        $this->feature = $feature;
-        $this->kind = $kind;
-    }
-
-    public static function unmarshal(array $json): ChairSearchCondition
-    {
-        return new ChairSearchCondition(
-            isset($json['width']) ? RangeCondition::unmarshal($json['width']) : null,
-            isset($json['height']) ? RangeCondition::unmarshal($json['height']) : null,
-            isset($json['depth']) ? RangeCondition::unmarshal($json['depth']) : null,
-            isset($json['price']) ? RangeCondition::unmarshal($json['price']) : null,
-            isset($json['color']) ? RangeCondition::unmarshal($json['color']) : null,
-            isset($json['features']) ? RangeCondition::unmarshal($json['features']) : null,
-            isset($json['kind']) ? RangeCondition::unmarshal($json['kind']) : null,
-        );
-    }
-}
-
-class EstateSearchCondition
-{
-    public ?RangeCondition $doorWidth;
-    public ?RangeCondition $doorHeight;
-    public ?RangeCondition $rent;
-    public ?ListCondition $feature;
-
-    public function __construct(
-        RangeCondition $doorHeight = null,
-        RangeCondition $doorWidth = null,
-        RangeCondition $rent = null,
-        ListCondition $feature = null
-    ) {
-        $this->doorHeight = $doorHeight;
-        $this->doorWidth = $doorWidth;
-        $this->rent = $rent;
-        $this->feature = $feature;
-    }
-
-    public static function unmarshal(array $json): EstateSearchCondition
-    {
-        return new EstateSearchCondition(
-            isset($json['doorHeight']) ? RangeCondition::unmarshal($json['doorHeight']) : null,
-            isset($json['doorWidth']) ? RangeCondition::unmarshal($json['doorWidth']) : null,
-            isset($json['rent']) ? RangeCondition::unmarshal($json['rent']) : null,
-            isset($json['feature']) ? ListCondition::unmarshal($json['feature']) : null,
-        );
-    }
-}
-
-class ListCondition
-{
-    /** @var string[] */
-    public array $list;
-
-    public function __construct(array $list)
-    {
-        $this->list = $list;
-    }
-
-    public static function unmarshal(array $list): ListCondition
-    {
-        return new ListCondition($list);
-    }
-}
 
 function getRange($condition, int $rangeId)
 {
     // $rangeId
 }
-
-global $chairSearchCondition;
-global $estateSearchCondition;
-
-function init() {
-    global $chairSearchCondition;
-    global $estateSearchCondition;
-
-    if (!$jsonText = file_get_contents(DIR_ROOT . '/fixture/chair_condition.json')) {
-        throw new RuntimeException(sprintf('Failed to get load file: %s', '/fixture/chair_condition.json'));
-    }
-    if (!$json = json_decode($jsonText, true)) {
-        throw new RuntimeException(sprintf('Failed to parse json: %s', '../fixture/chair_condition.json'));
-    }
-    $chairSearchCondition = ChairSearchCondition::unmarshal($json);
-
-    if (!$jsonText = file_get_contents(DIR_ROOT . '/fixture/estate_condition.json')) {
-        throw new RuntimeException(sprintf('Failed to get load file: %s', '/fixture/estate_condition.json'));
-    }
-    if (!$json = json_decode($jsonText, true)) {
-        throw new RuntimeException(sprintf('Failed to parse json: %s', '../fixture/estate_condition.json'));
-    }
-    $estateSearchCondition = EstateSearchCondition::unmarshal($json);
-}
-
-init();
 
 return function (App $app) {
     $app->options('/{routes:.*}', function (Request $request, Response $response) {
